# lib/api-zod/src/generated/types/ttsInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TtsInputVoice } from "./ttsInputVoice";

export interface TtsInput {
  text: string;
  voice?: TtsInputVoice;
}